+++ app/src/main/AndroidManifest.xml
@@ -14,8 +14,9 @@
         android:theme="@style/Theme.Owori"
         tools:targetApi="31">
         <activity
-            android:name=".ui.auth.view.AuthActivity"
+            android:name=".auth.ui.view.AuthActivity"
             android:exported="true"
+            android:screenOrientation="portrait"
             android:label="@string/app_name"
             android:theme="@style/Theme.Owori">
             <intent-filter>
