+++ app/src/main/java/com/owori/android/auth/ui/view/AuthActivity.kt
@@ -1,0 +1,30 @@
+package com.owori.android.auth.ui.view
+
+import androidx.navigation.fragment.NavHostFragment
+import androidx.navigation.fragment.findNavController
+import com.owori.android.common.ui.view.BaseActivity
+import com.owori.android.R
+import com.owori.android.databinding.ActivityAuthBinding
+import com.owori.android.auth.ui.viewmodel.AuthViewModel
+import org.koin.android.ext.android.inject
+
+class AuthActivity : BaseActivity<ActivityAuthBinding, AuthViewModel>(R.layout.activity_auth) {
+    override val viewModel: AuthViewModel by inject()
+
+    override fun initView() {
+        initNavGraph()
+    }
+
+    override fun initObserver() {}
+
+    override fun setBindingVariables(binding: ActivityAuthBinding) {}
+
+    private fun initNavGraph() {
+        val navHostFragment =
+            supportFragmentManager.findFragmentById(R.id.fragment_container) as NavHostFragment
+        val navController = navHostFragment.findNavController()
+        navController.navInflater.inflate(R.navigation.auth_navigation).apply {
+            setStartDestination(R.id.SplashFragment)
+        }.run { navController.setGraph(this, null) }
+    }
+}
