+++ app/src/main/java/com/owori/android/auth/ui/view/OnBoardingItemFragment.kt
@@ -1,0 +1,13 @@
+package com.owori.android.auth.ui.view
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+
+class OnBoardingItemFragment(private val fragmentId: Int) : Fragment() {
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+        return inflater.inflate(fragmentId, container, false)
+    }
+}
