+++ app/src/main/java/com/owori/android/auth/ui/viewmodel/PolicyViewModel.kt
@@ -1,0 +1,5 @@
+package com.owori.android.auth.ui.viewmodel
+
+import com.owori.android.common.ui.viewmodel.BaseViewModel
+
+class PolicyViewModel : BaseViewModel() {}
