+++ app/src/main/java/com/owori/android/auth/ui/viewmodel/SplashViewModel.kt
@@ -1,0 +1,5 @@
+package com.owori.android.auth.ui.viewmodel
+
+import com.owori.android.common.ui.viewmodel.BaseViewModel
+
+class SplashViewModel : BaseViewModel() {}
