+++ app/src/main/java/com/owori/android/common/AppConstants.kt
@@ -1,0 +1,5 @@
+package com.owori.android.common
+
+object AppConstants {
+
+}
