+++ app/src/main/java/com/owori/android/common/FragmentExt.kt
@@ -1,0 +1,8 @@
+package com.owori.android.common
+
+import androidx.fragment.app.Fragment
+import androidx.navigation.fragment.findNavController
+
+fun Fragment.navigateTo(res: Int) {
+    findNavController().navigate(res)
+}
