--- app/src/main/java/com/owrori/android/ui/OworiApplication.kt
@@ -1,15 +1,0 @@
-package com.owrori.android
-
-import android.app.Application
-import org.koin.android.ext.koin.androidContext
-import org.koin.core.context.startKoin
-
-class OworiApplication: Application() {
-    override fun onCreate() {
-        super.onCreate()
-        startKoin {
-            androidContext(this@OworiApplication)
-            modules(appModules)
-        }
-    }
-}
