--- app/src/main/java/com/owrori/android/ui/auth/viewmodel/OnBoardingViewModel.kt
@@ -1,7 +1,0 @@
-package com.owrori.android.ui.auth.viewmodel
-
-import com.owrori.android.ui.common.presentation.viewmodel.BaseViewModel
-
-class OnBoardingViewModel : BaseViewModel() {
-
-}
