--- app/src/main/java/com/owrori/android/ui/auth/viewmodel/SplashViewModel.kt
@@ -1,5 +1,0 @@
-package com.owrori.android.ui.auth.viewmodel
-
-import com.owrori.android.ui.common.presentation.viewmodel.BaseViewModel
-
-class SplashViewModel : BaseViewModel() {}
