--- app/src/main/java/com/owrori/android/ui/common/AppConstants.kt
@@ -1,5 +1,0 @@
-package com.owrori.android.ui.common
-
-object AppConstants {
-
-}
