--- app/src/main/java/com/owrori/android/ui/common/FragmentExt.kt
@@ -1,8 +1,0 @@
-package com.owrori.android.ui.common
-
-import androidx.fragment.app.Fragment
-import androidx.navigation.fragment.findNavController
-
-fun Fragment.navigateTo(res: Int) {
-    findNavController().navigate(res)
-}
