--- app/src/main/java/com/owrori/android/ui/common/SingleLiveEvent.kt
@@ -1,43 +1,0 @@
-package com.owrori.android.ui.common
-
-import android.util.Log
-import androidx.annotation.MainThread
-import androidx.lifecycle.LifecycleOwner
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.Observer
-import java.util.concurrent.atomic.AtomicBoolean
-
-open class SingleLiveEvent<T> : MutableLiveData<T>() {
-
-    private val mPending = AtomicBoolean(false)
-
-    override fun observe(owner: LifecycleOwner, observer: Observer<in T>) {
-        if (hasActiveObservers()) {
-            Log.w(TAG, "Multiple observers registered but only one will be notified of changes.")
-        }
-
-        super.observe(owner) { t ->
-            if (mPending.compareAndSet(true, false)) {
-                observer.onChanged(t)
-            }
-        }
-    }
-
-    @MainThread
-    override fun setValue(t: T?) {
-        mPending.set(true)
-        super.setValue(t)
-    }
-
-    /**
-     * Used for cases where T is Void, to make calls cleaner.
-     */
-    @MainThread
-    fun call() {
-        value = null
-    }
-
-    companion object {
-        private const val TAG = "SingleLiveEvent"
-    }
-}
