+++ app/src/main/res/drawable/kakao_login_icon.xml
@@ -1,0 +1,16 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="18dp"
+    android:height="18dp"
+    android:viewportWidth="18"
+    android:viewportHeight="18">
+  <group>
+    <clip-path
+        android:pathData="M0,0h18v18h-18z"/>
+    <path
+        android:pathData="M9,0.944C4.29,0.944 0,4.73 0,7.933C0,10.333 1.558,12.45 3.931,13.708L2.933,17.374C2.844,17.699 3.213,17.957 3.496,17.77L7.873,14.865C8.242,14.901 8.618,14.922 9,14.922C13.97,14.922 18,11.793 18,7.933C18,4.73 13.97,0.944 9,0.944Z"
+        android:strokeAlpha="0.902"
+        android:fillColor="#000000"
+        android:fillType="evenOdd"
+        android:fillAlpha="0.902"/>
+  </group>
+</vector>
