+++ app/src/main/res/drawable/owori_login_logo.xml
@@ -1,0 +1,42 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="207dp"
+    android:height="307dp"
+    android:viewportWidth="207"
+    android:viewportHeight="307">
+  <path
+      android:pathData="M80.38,60.18C80.38,93.42 53.44,120.36 20.2,120.36C-13.03,120.36 -39.97,93.42 -39.97,60.18C-39.97,26.94 -13.03,0 20.2,0C53.44,0 80.38,26.94 80.38,60.18Z"
+      android:fillColor="#E9E9E9"/>
+  <path
+      android:pathData="M-30.49,54.53C-1.66,71.13 8.22,107.9 -8.42,136.65C-25.07,165.41 -61.95,175.26 -90.78,158.66C-119.62,142.06 -129.5,105.29 -112.85,76.54C-96.2,47.78 -59.33,37.93 -30.49,54.53Z"
+      android:fillColor="#E9E9E9"/>
+  <path
+      android:pathData="M131.16,158.66C102.32,175.26 65.45,165.41 48.8,136.65C32.15,107.9 42.03,71.13 70.87,54.53C99.7,37.93 136.58,47.78 153.23,76.54C169.88,105.29 160,142.06 131.16,158.66Z"
+      android:fillColor="#E9E9E9"/>
+  <path
+      android:pathData="M80.38,246.07C80.38,279.3 53.44,306.25 20.2,306.25C-13.03,306.25 -39.97,279.3 -39.97,246.07C-39.97,212.83 -13.03,185.88 20.2,185.88C53.44,185.88 80.38,212.83 80.38,246.07Z"
+      android:fillColor="#E9E9E9"/>
+  <path
+      android:pathData="M131.16,147.59C159.99,164.19 169.88,200.96 153.23,229.71C136.58,258.47 99.7,268.32 70.87,251.72C42.03,235.12 32.15,198.35 48.8,169.6C65.45,140.84 102.32,130.99 131.16,147.59Z"
+      android:fillColor="#E9E9E9"/>
+  <path
+      android:pathData="M-30.49,251.72C-59.33,268.32 -96.2,258.47 -112.85,229.71C-129.5,200.96 -119.62,164.19 -90.79,147.59C-61.95,130.99 -25.07,140.84 -8.42,169.6C8.22,198.35 -1.66,235.12 -30.49,251.72Z"
+      android:fillColor="#E9E9E9"/>
+  <path
+      android:pathData="M104,60.18C104,93.42 77.06,120.36 43.83,120.36C10.6,120.36 -16.34,93.42 -16.34,60.18C-16.34,26.94 10.6,0 43.83,0C77.06,0 104,26.94 104,60.18Z"
+      android:fillColor="#FA7B53"/>
+  <path
+      android:pathData="M-6.87,54.53C21.97,71.13 31.85,107.9 15.2,136.65C-1.45,165.41 -38.32,175.26 -67.16,158.66C-96,142.06 -105.88,105.29 -89.23,76.54C-72.58,47.78 -35.7,37.93 -6.87,54.53Z"
+      android:fillColor="#FA7B53"/>
+  <path
+      android:pathData="M154.79,158.66C125.95,175.26 89.07,165.41 72.42,136.65C55.78,107.9 65.66,71.13 94.49,54.53C123.33,37.93 160.21,47.78 176.85,76.54C193.5,105.29 183.62,142.06 154.79,158.66Z"
+      android:fillColor="#FA7B53"/>
+  <path
+      android:pathData="M104,246.07C104,279.3 77.06,306.25 43.83,306.25C10.6,306.25 -16.34,279.3 -16.34,246.07C-16.34,212.83 10.6,185.88 43.83,185.88C77.06,185.88 104,212.83 104,246.07Z"
+      android:fillColor="#FA7B53"/>
+  <path
+      android:pathData="M154.79,147.59C183.62,164.19 193.5,200.96 176.85,229.71C160.2,258.47 123.33,268.32 94.49,251.72C65.66,235.12 55.78,198.35 72.42,169.6C89.07,140.84 125.95,130.99 154.79,147.59Z"
+      android:fillColor="#FA7B53"/>
+  <path
+      android:pathData="M-6.87,251.72C-35.7,268.32 -72.58,258.47 -89.23,229.71C-105.88,200.96 -96,164.19 -67.16,147.59C-38.32,130.99 -1.45,140.84 15.2,169.6C31.85,198.35 21.97,235.12 -6.87,251.72Z"
+      android:fillColor="#FA7B53"/>
+</vector>
