+++ app/src/main/res/drawable/owori_login_typo.xml
@@ -1,0 +1,15 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="123dp"
+    android:height="52dp"
+    android:viewportWidth="123"
+    android:viewportHeight="52">
+  <path
+      android:pathData="M18.5,0.63C30.9,0.6 37.1,6.57 37.1,18.53C37.07,28.9 32.02,34.62 21.95,35.68V42.43H35.15C36.45,42.43 37.1,43.17 37.1,44.63V48.03C37.1,49.27 36.48,49.9 35.25,49.93H2.3C0.8,49.93 0.05,49.17 0.05,47.63V44.63C0.05,43.17 0.85,42.43 2.45,42.43H14V35.58C4.7,34.18 0.03,28.5 0,18.53C0.03,6.6 6.2,0.63 18.5,0.63ZM18.45,29.08C25.58,29.08 29.25,25.57 29.45,18.53C29.45,11.1 25.8,7.42 18.5,7.48C11.37,7.48 7.77,11.17 7.7,18.53C7.77,25.67 11.35,29.18 18.45,29.08Z"
+      android:fillColor="#000000"/>
+  <path
+      android:pathData="M44.73,36.53C43.39,36.53 42.73,35.97 42.73,34.83V32.88C42.73,31.55 43.34,30.88 44.58,30.88H52.03V28.48H44.78C43.34,28.48 42.63,27.77 42.63,26.33V23.73C42.63,22.3 43.33,21.58 44.73,21.58H67.53C68.73,21.62 69.33,22.27 69.33,23.53V26.33C69.33,27.77 68.68,28.48 67.38,28.48H59.88V30.88H77.48C78.88,30.88 79.58,31.48 79.58,32.68V41.53C79.58,42.73 78.86,43.33 77.43,43.33H50.33V45.33H77.58C78.91,45.33 79.58,45.9 79.58,47.03V48.83C79.58,50.27 78.96,50.98 77.73,50.98H45.03C43.49,50.98 42.73,50.4 42.73,49.23V40.33C42.73,39.13 43.44,38.53 44.88,38.53H71.88V36.53H44.73ZM55.43,0.58C62.39,0.58 66.49,2.92 67.73,7.58H71.48V2.88C71.48,1.72 72.13,1.13 73.43,1.13H77.58C78.91,1.13 79.58,1.72 79.58,2.88V27.53C79.58,28.73 78.89,29.33 77.53,29.33H73.53C72.16,29.33 71.48,28.7 71.48,27.43V15.08H67.23C65.59,18.52 61.64,20.23 55.38,20.23C47.08,20.17 42.86,17 42.73,10.73C42.79,4 47.03,0.62 55.43,0.58ZM55.53,13.43C58.96,13.43 60.68,12.53 60.68,10.73C60.68,8.53 58.99,7.43 55.63,7.43C52.23,7.43 50.53,8.5 50.53,10.63C50.53,12.53 52.19,13.47 55.53,13.43Z"
+      android:fillColor="#000000"/>
+  <path
+      android:pathData="M122.11,49.23C122.11,50.43 121.51,51.03 120.31,51.03H116.01C114.78,51.03 114.16,50.38 114.16,49.08V2.88C114.16,1.72 114.78,1.13 116.01,1.13H120.16C121.46,1.13 122.11,1.75 122.11,2.98V49.23ZM98.06,1.13C100.66,1.13 102.79,1.63 104.46,2.63C106.13,3.63 107.44,5.15 108.41,7.18C109.38,9.18 110.04,11.7 110.41,14.73C110.78,17.73 110.96,21.25 110.96,25.28C110.96,29.58 110.76,33.35 110.36,36.58C109.96,39.78 109.28,42.47 108.31,44.63C107.34,46.77 106.03,48.37 104.36,49.43C102.69,50.5 100.59,51.03 98.06,51.03C95.46,51.03 93.31,50.48 91.61,49.38C89.91,48.32 88.56,46.72 87.56,44.58C86.59,42.42 85.91,39.73 85.51,36.53C85.11,33.3 84.91,29.55 84.91,25.28C84.91,21.28 85.13,17.78 85.56,14.78C86.03,11.75 86.76,9.23 87.76,7.23C88.79,5.2 90.14,3.68 91.81,2.68C93.51,1.65 95.59,1.13 98.06,1.13ZM103.46,25.28C103.46,22.45 103.39,19.97 103.26,17.83C103.13,15.7 102.86,13.93 102.46,12.53C102.06,11.1 101.51,10.03 100.81,9.33C100.11,8.63 99.21,8.28 98.11,8.28C96.98,8.28 96.06,8.63 95.36,9.33C94.66,10.03 94.11,11.1 93.71,12.53C93.34,13.93 93.09,15.7 92.96,17.83C92.83,19.97 92.76,22.45 92.76,25.28C92.76,31.52 93.13,36.2 93.86,39.33C94.59,42.43 96.01,43.98 98.11,43.98C99.18,43.98 100.06,43.6 100.76,42.83C101.46,42.03 102.01,40.85 102.41,39.28C102.81,37.72 103.08,35.77 103.21,33.43C103.38,31.1 103.46,28.38 103.46,25.28Z"
+      android:fillColor="#000000"/>
+</vector>
